requirements.txt
flask~=3.0.3
python-consul2
requests~=2.32.3